Update graphic.php
Changed:

    if($_GET['type'] == background)

to:

    if(strpos($_SERVER['REQUEST_URI'],'background') == true)

because it wasn't getting the type from $_GET for some reason. This new line has it look in the URL for the word 'background' and if it sees it it assumes it's a background and if it doesn't it assumes it's an avatar. This can probably be fixed in .htaccess later on but it is a good fix for the time being.